+++ CHANGES.rst
@@ -8,8 +8,13 @@ Changes
 
 Changes:
 --------
-- Add AWS S3 bucket support for process inputs references.
+- Add AWS S3 bucket support for process input references.
 - Add ``weaver.wps_output_bucket`` configuration to upload results to AWS S3 bucket instead of local directory.
+- Add more documentation about supported file reference schemes.
+- Add documentation references to `ESGF-CWT Compute API`.
+- Add conditional input file reference fetching (depending on `ADES`/`EMS`, process *type*  from `CWL` ``hints``)
+  to take advantage of *request-options* and all supported scheme formats by `Weaver`, instead of relying on ``PyWPS``
+  and/or ``CWL`` wherever how far downstream the URL reference was reaching.
 
 Fixes:
 ------
@@ -20,8 +25,8 @@ Fixes:
 
 Changes:
 --------
-- Add multiple CWL ESGF processes and workflows, namely ``SubsetNASAESGF``, ``SubsetNASAESGF`` and many more.
-- Add tests for ESGF processes and workflows.
+- Add multiple `CWL` `ESGF` processes and workflows, namely ``SubsetNASAESGF``, ``SubsetNASAESGF`` and many more.
+- Add tests for `ESGF` processes and workflows.
 - Add documentation for ``ESGF-CWTRequirement`` processes.
 - Add ``file2string_array`` and ``metalink2netcdf`` builtins.
 - Add ``esgf_process`` ``Wps1Process`` extension, to handle ``ESGF-CWTRequirement`` processes and workflows.